feat-chore: action type apis rearrange
