+++ autobots/api/v1/api_actions/img2img/image_mixer_stable_diffusion.py
